// src/services/product/models/AddProductDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BrandDTO } from './BrandDTO';
export type AddProductDTO = {
    /**
     * product name
     */
    name?: string;
    brand?: BrandDTO;
    /**
     * short description of product
     */
    description?: string;
};